fix: move preferences back to profile and add transitions for badges
