+++ desk/src/components/Settings/FieldDependency/FieldDependency.vue
@@ -10,7 +10,11 @@
           @click="handleBackNavigation"
           class="cursor-pointer -ml-4 hover:bg-transparent focus:bg-transparent focus:outline-none focus:ring-0 focus:ring-offset-0 focus-visible:none active:bg-transparent active:outline-none active:ring-0 active:ring-offset-0 active:text-ink-gray-5 font-semibold text-ink-gray-7 text-lg hover:opacity-70 !pr-0"
         />
-        <Badge v-if="isDirty" theme="orange"> {{ __("Unsaved") }} </Badge>
+        <Transition name="fade">
+          <Badge v-if="isDirty" theme="orange">
+            {{ __("Unsaved") }}
+          </Badge></Transition
+        >
       </div>
     </template>
     <template #header-actions>
