+++ desk/src/components/Settings/Sla/SlaPolicyView.vue
@@ -10,13 +10,14 @@
           @click="goBack()"
           class="cursor-pointer -ml-4 hover:bg-transparent focus:bg-transparent focus:outline-none focus:ring-0 focus:ring-offset-0 focus-visible:none active:bg-transparent active:outline-none active:ring-0 active:ring-offset-0 active:text-ink-gray-5 font-semibold text-ink-gray-7 text-lg hover:opacity-70 !pr-0"
         />
-        <Badge
-          :variant="'subtle'"
-          :theme="'orange'"
-          size="sm"
-          :label="__('Unsaved')"
-          v-if="isDirty"
-        />
+        <Transition name="fade">
+          <Badge
+            :variant="'subtle'"
+            :theme="'orange'"
+            size="sm"
+            :label="__('Unsaved')"
+            v-if="isDirty"
+        /></Transition>
       </div>
     </template>
     <template #header-actions>
