--- desk/src/components/icons/SlidersIcon.vue
@@ -1,14 +1,0 @@
-<template>
-  <svg
-    width="16"
-    height="16"
-    viewBox="0 0 16 16"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M3 8.5C4.10457 8.5 5 9.39543 5 10.5C5 11.4316 4.3622 12.2113 3.5 12.4336V14C3.5 14.2761 3.27614 14.5 3 14.5C2.72386 14.5 2.5 14.2761 2.5 14V12.4336C1.6378 12.2113 1 11.4316 1 10.5C1 9.39543 1.89543 8.5 3 8.5ZM8 9.5C8.27614 9.5 8.5 9.72386 8.5 10V14C8.5 14.2761 8.27614 14.5 8 14.5C7.72386 14.5 7.5 14.2761 7.5 14V10C7.5 9.72386 7.72386 9.5 8 9.5ZM13 8.5C14.1046 8.5 15 9.39543 15 10.5C15 11.4316 14.3622 12.2113 13.5 12.4336V14C13.5 14.2761 13.2761 14.5 13 14.5C12.7239 14.5 12.5 14.2761 12.5 14V12.4336C11.6378 12.2113 11 11.4316 11 10.5C11 9.39543 11.8954 8.5 13 8.5ZM3 9.5C2.44772 9.5 2 9.94772 2 10.5C2 11.0523 2.44772 11.5 3 11.5C3.55228 11.5 4 11.0523 4 10.5C4 9.94772 3.55228 9.5 3 9.5ZM13 9.5C12.4477 9.5 12 9.94772 12 10.5C12 11.0523 12.4477 11.5 13 11.5C13.5523 11.5 14 11.0523 14 10.5C14 9.94772 13.5523 9.5 13 9.5ZM8 2C8.27614 2 8.5 2.22386 8.5 2.5V4.5C8.5 4.52163 8.49583 4.54257 8.49316 4.56348C9.35898 4.7833 10 5.56581 10 6.5C10 7.60457 9.10457 8.5 8 8.5C6.89543 8.5 6 7.60457 6 6.5C6 5.56614 6.64051 4.78362 7.50586 4.56348C7.50321 4.54261 7.5 4.52158 7.5 4.5V2.5C7.5 2.22386 7.72386 2 8 2ZM8 5.5C7.44772 5.5 7 5.94772 7 6.5C7 7.05228 7.44772 7.5 8 7.5C8.55228 7.5 9 7.05228 9 6.5C9 5.94772 8.55228 5.5 8 5.5ZM3 2C3.27614 2 3.5 2.22386 3.5 2.5V6.5C3.5 6.77614 3.27614 7 3 7C2.72386 7 2.5 6.77614 2.5 6.5V2.5C2.5 2.22386 2.72386 2 3 2ZM13 2C13.2761 2 13.5 2.22386 13.5 2.5V6.5C13.5 6.77614 13.2761 7 13 7C12.7239 7 12.5 6.77614 12.5 6.5V2.5C12.5 2.22386 12.7239 2 13 2Z"
-      fill="currentColor"
-    />
-  </svg>
-</template>
