+++ desk/src/index.css
@@ -46,3 +46,13 @@ body {
 *::-webkit-scrollbar-corner {
   background: transparent;
 }
+
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.2s ease;
+}
+
+.fade-enter-from,
+.fade-leave-to {
+  opacity: 0;
+}
